ud-03 ejs.

+++ ud-01-sintaxis-basica/e3.py
@@ -1,4 +1,4 @@
-lado = int(input("Introduzca el area del cuadrado: "))
+lado = float(input("Introduzca el area del cuadrado: "))
 
 area = lado ** 2
